+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix data race in `TestRemotelyControlledSampler_updateRateLimitingOrProbabilisticSampler` in `go.opentelemetry.io/contrib/samplers/jaegerremote`.
+
 <!-- Released section -->
 <!-- Don't change this section unless doing release -->